requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgi-redis==0.14.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.11.2
Automat==0.7.0
Babel==2.6.0
billiard==3.3.0.23
celery==3.1.23
certifi==2018.10.15
channels==2.1.5
chardet==3.0.4
constantly==15.1.0
daphne==2.2.3
defusedxml==0.5.0
Django==2.1.3
django-bootstrap3==11.0.0
django-braces==1.13.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.9.0
flower==0.9.2
gunicorn==19.9.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
kombu==3.0.35
msgpack-python==0.4.7
oauthlib==2.1.0
pkg-resources==0.0.0
PyHamcrest==1.9.0
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.7
redis==2.10.5
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
tornado==5.1.1
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
vine==1.1.4
whitenoise==4.1.2
zope.interface==4.6.0
